third_party/blink/web_tests/external/wpt/css/cssom-view/scrollParent-expected.txt

This is a testharness.js-based test.
[FAIL] scrollParent returns the nearest scroll container.
  normalChild.scrollParent is not a function
[FAIL] hidden element is a scroll container.
  childOfHidden.scrollParent is not a function
[FAIL] Element with no box has null scrollParent.
  noBox.scrollParent is not a function
[FAIL] scrollParent follows absolute positioned containing block chain.
  absPosChild.scrollParent is not a function
[FAIL] scrollParent follows fixed positioned containing block chain.
  fixedPosChild.scrollParent is not a function
[FAIL] scrollParent of element fixed to root is null.
  fixedToRoot.scrollParent is not a function
[FAIL] scrollParent of child in root viewport returns document scrolling element.
  childOfRoot.scrollParent is not a function
[FAIL] scrollParent of fixed element contained within root is document scrolling element.
  fixedContainedByRoot.scrollParent is not a function
[FAIL] scrollParent of body is null.
  document.body.scrollParent is not a function
[FAIL] scrollParent of root is null.
  document.documentElement.scrollParent is not a function
[FAIL] scrollParent skips ancestors with `display: contents`.
  childOfDisplayContents.scrollParent is not a function
Harness: the test ran to completion.
